+++ packages/react/src/components/toast/toast.stories.tsx
@@ -1,8 +1,8 @@
-import React from "react";
-import { Toast } from ".";
-import { Card } from "../card";
-import { Button } from "../button";
 import type { Story, StoryDefault } from "@ladle/react";
+import React from "react";
+import { Button } from "../button";
+import { Card } from "../card";
+import { Toast } from ".";
 
 export default {
 	title: "Toast",
@@ -35,7 +35,7 @@ export const Base: Story<Toast.RootProps> = () => {
       <Toast.Group toaster={toaster}>
         {(toast) => (
           <Toast.Root key={toast.id} asChild>
-            <Card.Root visual="elevated" radius="2xl" gap="12" layout="inline" py="12" px="16" alignItems="center">
+            <Card.Root visual="elevated" radius="2xl" gap="12" orientation="horizontal" py="12" px="16" alignItems="center">
               <Card.Body>
                 <Toast.Title>
                   {toast.title}